Move LinkContext to /context folder. Rename 'p' variant to 'default'. Create TextElements type
