+++ libs/island-ui/core/src/lib/context/index.ts
@@ -1,1 +1,2 @@
 export * from './ColorSchemeContext/ColorSchemeContext'
+export { LinkContext } from './LinkContext'
